+++ client/Packages/com.beamable/Common/Runtime/Api/Auth/AuthApi.cs
@@ -2,7 +2,6 @@ using Beamable.Common.Content;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Web;
 using UnityEngine;
 
 namespace Beamable.Common.Api.Auth
@@ -52,10 +51,7 @@ namespace Beamable.Common.Api.Auth
 
 		public Promise<bool> IsThirdPartyAvailable(AuthThirdParty thirdParty, string token)
 		{
-			var query = HttpUtility.ParseQueryString(string.Empty);
-			query["thirdParty"] = thirdParty.GetString();
-			query["token"] = token;
-
+			string query = _requester.EscapeURL($"thirdParty={thirdParty.GetString()}&token={token}");
 			return _requester
 				   .Request<AvailabilityResponse>(
 					   Method.GET,
@@ -150,10 +146,7 @@ namespace Beamable.Common.Api.Auth
 
 		public Promise<User> RemoveThirdPartyAssociation(AuthThirdParty thirdParty, string token)
 		{
-			var query = HttpUtility.ParseQueryString(string.Empty);
-			query["thirdParty"] = thirdParty.GetString();
-			query["token"] = token;
-
+			string query = _requester.EscapeURL($"thirdParty={thirdParty.GetString()}&token={token}");
 			return _requester.Request<User>(Method.DELETE,
 											$"{ACCOUNT_URL}/me/third-party?{query}",
 											null, true);
@@ -405,11 +398,10 @@ namespace Beamable.Common.Api.Auth
 
 		public Promise<bool> IsExternalIdentityAvailable(string providerService, string externalToken, string providerNamespace = null)
 		{
-			var query = HttpUtility.ParseQueryString(string.Empty);
-			query["user_id"] = externalToken;
-			query["provider_service"] = providerService;
-			if (!string.IsNullOrEmpty(providerNamespace))
-				query["provider_namespace"] = providerNamespace;
+			string query = _requester.EscapeURL($"provider_service={providerService}&user_id={externalToken}");
+
+			if (!string.IsNullOrWhiteSpace(providerNamespace))
+				query += _requester.EscapeURL($"&{providerNamespace}");
 
 			return Requester.Request<AvailabilityResponse>(
 				Method.GET,